# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ระบบจัดการการลาของพนักงาน — Backend API สำหรับยื่นใบลา อนุมัติ ปฏิเสธ และดูยอดวันลาคงเหลือ",
        "title": "Leave Management System API",
        "contact": {
            "name": "API Support",
            "email": "support@company.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/leaves": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "สร้างคำขอลาใหม่ ระบบจะตรวจสอบวันลาซ้ำซ้อนและยอดวันลาคงเหลือโดยอัตโนมัติ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave"
                ],
                "summary": "ยื่นใบลาใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลสำหรับยื่นใบลา",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitLeaveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.LeaveRequestResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/leaves/my-balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลยอดวันลาคงเหลือทุกประเภทของผู้ใช้ที่เข้าสู่ระบบ",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave"
                ],
                "summary": "ดูยอดวันลาคงเหลือ",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.LeaveBalanceResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/leaves/my-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลใบลาทั้งหมดของผู้ใช้ที่เข้าสู่ระบบ (รองรับ pagination)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leave"
                ],
                "summary": "ดูประวัติใบลา",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "หน้าที่ต้องการ (เริ่มจาก 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "จำนวนรายการต่อหน้า (สูงสุด 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginatedAPIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.LeaveRequestResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/manager/pending-requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลใบลาทั้งหมดที่มีสถานะ pending สำหรับผู้จัดการ (รองรับ pagination)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "ดูใบลารอการอนุมัติ",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "หน้าที่ต้องการ (เริ่มจาก 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "จำนวนรายการต่อหน้า (สูงสุด 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginatedAPIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.LeaveRequestResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/manager/requests/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ผู้จัดการอนุมัติคำขอลา ระบบจะหักยอดวันลาของพนักงานโดยอัตโนมัติ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "อนุมัติใบลา",
                "parameters": [
                    {
                        "type": "string",
                        "description": "รหัสใบลา (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "หมายเหตุจากผู้อนุมัติ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReviewLeaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/manager/requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ผู้จัดการปฏิเสธคำขอลา ยอดวันลาของพนักงานจะไม่เปลี่ยนแปลง",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manager"
                ],
                "summary": "ปฏิเสธใบลา",
                "parameters": [
                    {
                        "type": "string",
                        "description": "รหัสใบลา (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "หมายเหตุจากผู้อนุมัติ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReviewLeaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูลที่ส่งกลับ (ถ้ามี)"
                },
                "message": {
                    "description": "ข้อความอธิบาย",
                    "type": "string"
                },
                "success": {
                    "description": "สถานะความสำเร็จ",
                    "type": "boolean"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "รายละเอียดข้อผิดพลาด (ถ้ามี)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "ข้อความอธิบายข้อผิดพลาด",
                    "type": "string"
                },
                "success": {
                    "description": "สถานะ (false เสมอ)",
                    "type": "boolean"
                }
            }
        },
        "dto.LeaveBalanceResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "รหัสยอดวันลา",
                    "type": "string"
                },
                "leave_type": {
                    "description": "ประเภทการลา",
                    "type": "string"
                },
                "remaining_days": {
                    "description": "วันลาคงเหลือ",
                    "type": "number"
                },
                "total_days": {
                    "description": "วันลาทั้งหมด",
                    "type": "number"
                },
                "used_days": {
                    "description": "วันลาที่ใช้ไป",
                    "type": "number"
                },
                "year": {
                    "description": "ปี",
                    "type": "integer"
                }
            }
        },
        "dto.LeaveRequestResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "วันที่ยื่นใบลา",
                    "type": "string"
                },
                "end_date": {
                    "description": "วันสิ้นสุด",
                    "type": "string"
                },
                "id": {
                    "description": "รหัสใบลา",
                    "type": "string"
                },
                "leave_type": {
                    "description": "ประเภทการลา",
                    "type": "string"
                },
                "reason": {
                    "description": "เหตุผลการลา",
                    "type": "string"
                },
                "review_note": {
                    "description": "หมายเหตุจากผู้อนุมัติ",
                    "type": "string"
                },
                "reviewed_at": {
                    "description": "วันที่อนุมัติ/ปฏิเสธ",
                    "type": "string"
                },
                "reviewer_id": {
                    "description": "รหัสผู้อนุมัติ",
                    "type": "string"
                },
                "start_date": {
                    "description": "วันเริ่มต้น",
                    "type": "string"
                },
                "status": {
                    "description": "สถานะ (pending/approved/rejected)",
                    "type": "string"
                },
                "total_days": {
                    "description": "จำนวนวันลาทั้งหมด",
                    "type": "number"
                },
                "updated_at": {
                    "description": "วันที่แก้ไขล่าสุด",
                    "type": "string"
                },
                "user_id": {
                    "description": "รหัสพนักงาน",
                    "type": "string"
                }
            }
        },
...
        "dto.PaginatedAPIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "รายการข้อมูล"
                },
                "message": {
                    "description": "ข้อความอธิบาย",
                    "type": "string"
                },
                "pagination": {
                    "description": "ข้อมูล pagination",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.PaginationMeta"
                        }
                    ]
                },
                "success": {
                    "description": "สถานะความสำเร็จ",
                    "type": "boolean"
                }
            }
        },
        "dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "หน้าปัจจุบัน",
                    "type": "integer"
                },
                "page_size": {
                    "description": "จำนวนรายการต่อหน้า",
                    "type": "integer"
                },
                "total": {
                    "description": "จำนวนรายการทั้งหมด",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "จำนวนหน้าทั้งหมด",
                    "type": "integer"
                }
            }
        },
        "dto.ReviewLeaveRequest": {
            "type": "object",
            "properties": {
                "note": {
                    "description": "หมายเหตุจากผู้อนุมัติ (ไม่บังคับ)",
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
        "dto.SubmitLeaveRequest": {
            "type": "object",
            "required": [
                "end_date",
                "leave_type",
                "reason",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "description": "วันสิ้นสุด (YYYY-MM-DD)",
                    "type": "string"
                },
                "leave_type": {
                    "description": "ประเภทการลา",
                    "type": "string",
                    "enum": [
                        "sick_leave",
                        "annual_leave",
                        "personal_leave"
                    ]
                },
                "reason": {
                    "description": "เหตุผลการลา",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 5
                },
                "start_date": {
                    "description": "วันเริ่มต้น (YYYY-MM-DD)",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "กรุณาใส่ Bearer token เช่น \"Bearer eyJhbGciOiJIUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}